+++ python_modules/dagster/dagster/_core/instance/mixins/runs_mixin.py
@@ -10,14 +10,12 @@ from collections.abc import Mapping, Sequence, Set
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, Optional, Union, cast
 
-from dagster._annotations import public
 from dagster._core.definitions.events import AssetKey
 from dagster._core.storage.dagster_run import (
     DagsterRun,
     DagsterRunStatus,
     JobBucket,
     RunPartitionData,
-    RunRecord,
     RunsFilter,
     TagBucket,
 )
@@ -54,41 +52,6 @@ class RunsMixin:
     _event_storage: "EventLogStorage"
 
     # Core Run Retrieval Methods
-
-    @public
-    def get_run_by_id(self, run_id: str) -> Optional[DagsterRun]:
-        """Get a :py:class:`DagsterRun` matching the provided `run_id`.
-
-        Args:
-            run_id (str): The id of the run to retrieve.
-
-        Returns:
-            Optional[DagsterRun]: The run corresponding to the given id. If no run matching the id
-                is found, return `None`.
-        """
-        record = self.get_run_record_by_id(run_id)
-        if record is None:
-            return None
-        return record.dagster_run
-
-    @public
-    @traced
-    def get_run_record_by_id(self, run_id: str) -> Optional[RunRecord]:
-        """Get a :py:class:`RunRecord` matching the provided `run_id`.
-
-        Args:
-            run_id (str): The id of the run record to retrieve.
-
-        Returns:
-            Optional[RunRecord]: The run record corresponding to the given id. If no run matching
-                the id is found, return `None`.
-        """
-        if not run_id:
-            return None
-        records = self._run_storage.get_run_records(RunsFilter(run_ids=[run_id]), limit=1)
-        if not records:
-            return None
-        return records[0]
 
     @traced
     def get_run_stats(self, run_id: str) -> "DagsterRunStatsSnapshot":
@@ -297,48 +260,10 @@ class RunsMixin:
     def get_runs_count(self, filters: Optional[RunsFilter] = None) -> int:
         return self._run_storage.get_runs_count(filters)
 
-    @public
-    @traced
-    def get_run_records(
-        self,
-        filters: Optional[RunsFilter] = None,
-        limit: Optional[int] = None,
-        order_by: Optional[str] = None,
-        ascending: bool = False,
-        cursor: Optional[str] = None,
-        bucket_by: Optional[Union[JobBucket, TagBucket]] = None,
-    ) -> Sequence[RunRecord]:
-        """Return a list of run records stored in the run storage, sorted by the given column in given order.
-
-        Args:
-            filters (Optional[RunsFilter]): the filter by which to filter runs.
-            limit (Optional[int]): Number of results to get. Defaults to infinite.
-            order_by (Optional[str]): Name of the column to sort by. Defaults to id.
-            ascending (Optional[bool]): Sort the result in ascending order if True, descending
-                otherwise. Defaults to descending.
-
-        Returns:
-            List[RunRecord]: List of run records stored in the run storage.
-        """
-        return self._run_storage.get_run_records(
-            filters, limit, order_by, ascending, cursor, bucket_by
-        )
-
     @traced
     def get_run_partition_data(self, runs_filter: RunsFilter) -> Sequence[RunPartitionData]:
         """Get run partition data for a given partitioned job."""
         return self._run_storage.get_run_partition_data(runs_filter)
-
-    @public
-    @traced
-    def delete_run(self, run_id: str) -> None:
-        """Delete a run and all events generated by that from storage.
-
-        Args:
-            run_id (str): The id of the run to delete.
-        """
-        self._run_storage.delete_run(run_id)
-        self._event_storage.delete_events(run_id)
 
     @traced
     def get_records_for_run(
